backend/api/gateway/classifierAPI/requirements.txt
pydantic
pydantic_settings
fastapi
uvicorn
pymongo
pyjwt
joblib
tensorflow
pandas
kagglehub
numpy
scikit-learn
joblib
matplotlib